correct no boost reading of XML data
Just fixed some typos in the CXX side or the ifdef to match the
boost side. With these (and a slight increase in temporal tolerance)
the VTK xdmf tests pass.

And finally no BOOST! Kudos to Andrew for doing the hard work.

+++ XdmfGeometry.cpp
@@ -196,7 +196,7 @@ XdmfGeometry::populateItem(const std::map<std::string, std::string> & itemProper
     char * token = std::strtok(originString, " ");
     while (token != NULL)
     {
-      mOrigin.push_back(atoi(token));
+      mOrigin.push_back(atof(token));
       token = std::strtok(NULL, " ");
     }
